requirements.txt
glfw==2.5.5
Pillow==9.3.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
